requirements.txt
bertopic
sentence-transformers
umap-learn
hdbscan
pandas
numpy
scikit-learn